Instead two error files, I just made one and merge
