+++ src/components/buttons/lockButton.js
@@ -1,15 +1,26 @@
 import {
     ActionRowBuilder,
     EmbedBuilder,
+    PermissionFlagsBits,
     StringSelectMenuBuilder,
     StringSelectMenuOptionBuilder,
 } from "discord.js";
+import {
+    defaultBotPermError,
+    defaultUserPermError,
+} from "../../shortcuts/defaultPermissionsErrors.js";
 
 export default {
     data: {
         customId: "issue-lock-conversation",
     },
     execute: async ({ interaction }) => {
+        if (defaultBotPermError(interaction, PermissionFlagsBits.ManageThreads))
+            return;
+        if (
+            defaultUserPermError(interaction, PermissionFlagsBits.ManageThreads)
+        )
+            return;
         const lockButtonExplanation = new EmbedBuilder()
             .setTitle("Lock conversation on this issue")
             .setThumbnail(
